Make http_client module http_utils
