+++ lib/crawler/http_executor.rb
@@ -130,7 +130,7 @@ module Crawler
     #-------------------------------------------------------------------------------------------------
     # Returns an HTTP client to be used for all requests
     def http_client
-      @http_client ||= Crawler::HttpClient::Base.new(
+      @http_client ||= Crawler::HttpClient.new(
         pool_max: 100,
         user_agent: config.user_agent,
         loopback_allowed: config.loopback_allowed,
