+++ lib/crawler/http_utils/all_trusting_trust_manager.rb
@@ -5,7 +5,7 @@ java_import javax.net.ssl.X509TrustManager
 # A simple implementation of the trust manager interface that trusts everyone
 # Used by the Crawler HTTP client to implement ssl_verification_mode=none.
 module Crawler
-  module HttpClient
+  module HttpUtils
     class AllTrustingTrustManager
       include X509TrustManager
