+++ spec/lib/crawler/http_utils/response_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-RSpec.describe(Crawler::HttpClient::Response) do
+RSpec.describe(Crawler::HttpUtils::Response) do
   let(:url) { Crawler::Data::URL.parse('http://example.org/') }
   let(:response) do
-    Crawler::HttpClient::Response.new(
+    Crawler::HttpUtils::Response.new(
       apache_response: apache_response,
       url: url,
       request_start_time: 1.second.ago,
@@ -37,14 +37,14 @@ RSpec.describe(Crawler::HttpClient::Response) do
     context 'when given an unsupported content encoding' do
       let(:encoding) { 'banana' }
       it 'should fail' do
-        expect { check_content_encoding }.to raise_error(Crawler::HttpClient::InvalidEncoding)
+        expect { check_content_encoding }.to raise_error(Crawler::HttpUtils::InvalidEncoding)
       end
     end
 
     context 'when given a list with an unsupported content encoding' do
       let(:encoding) { 'gzip,banana' }
       it 'should fail' do
-        expect { check_content_encoding }.to raise_error(Crawler::HttpClient::InvalidEncoding)
+        expect { check_content_encoding }.to raise_error(Crawler::HttpUtils::InvalidEncoding)
       end
     end
   end
